Refactoring cluster.go
